software: reorganize module hierarchy. NFCI
Move all simulation-related implementation details to `.simulation`,
all hardware-related ones to `.hardware`, and remaining abstract base
classes to `.device` and `.access`.
