+++ software/glasgow/hardware/quirks.py
